LayoutTests/imported/w3c/web-platform-tests/feature-policy/feature-policy-frame-policy-timing.https.sub-expected.txt



FAIL allow attr timing test same origin assert_equals: expected "NotAllowedError" but got "ok"
FAIL allow attr timing test diff origin assert_equals: expected "NotAllowedError" but got "ok"
